+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [@lazy-promise/core@0.0.22](https://github.com/lazy-promise/lazy-promise/tree/%40lazy-promise/core%400.0.22)
+
+- `catchError` renamed to `catchRejection`.
+
 ## [@lazy-promise/core@0.0.21](https://github.com/lazy-promise/lazy-promise/tree/%40lazy-promise/core%400.0.21)
 
 - LazyPromise no longer stores and multicasts the result.